extracted transaction list
